extract asm statement parser helpers
